+++ src/variables/BlitzVariable.php
@@ -220,7 +220,7 @@ class BlitzVariable
         ];
 
         if ($config->requestType === VariableConfigModel::INCLUDE_REQUEST_TYPE) {
-            if (Craft::$app->getRequest()->getIsPreview() || Craft::$app->getRequest()->getIsLivePreview()) {
+            if (Blitz::$cacheRequest->getIsPreviewOrTokenRequest()) {
                 return Template::raw(Craft::$app->getView()->renderTemplate($template, $params));
             }